+++ test.js
@@ -1,4 +1,4 @@
-const { unicode, emojis, cats, hearts, circles } = require('./index.js');
+const { unicode, emojis, cats, hearts, circles, kaomojis } = require('./index.js');
 
 // A single emote
 console.log(unicode());
@@ -18,3 +18,13 @@ console.log(`Name: ${heart.name}; Emoji: ${heart.content};`);
 // Random circles
 const circle = circles();
 console.log(`Name: ${circle.name}; Emoji: ${circle.content};`);
+
+// Random kaomoji
+const kaomoji = new kaomojis();
+
+async function example() {
+	const uwu = await kaomoji.uwu();
+	console.log(uwu);
+}
+
+example();