+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 📥・v1.2.0 (12.05.2022)
 1. Updated CHANGELOG.md, README.md and package.json.
-2. Added new function `circle()` with random cat hearts.
+2. Added new function `circle()` with random circles.
 
 ## 📥・v1.1.1 (11.05.2022)
-1. Added new function `hearts()` with random cat hearts.
+1. Added new function `hearts()` with random hearts.
 
 ## 📥・v1.0.1 (08.05.2022)
 1. Added new function `cats()` with random cat emojis.