+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 📥・v1.3.0 (13.05.2022)
+1. Updated README.md file.
+2. Added new function `kaomojis()` with random kaomoji from my https://api.skiffybot.pl API.
+
 ## 📥・v1.2.0 (12.05.2022)
 1. Updated CHANGELOG.md, README.md and package.json.
 2. Added new function `circles()` with random circles.